Optimize filtering and lazy-load heavy profile/job dialogs
